refactor: 요청 DTO를 Zod 스키마로 전환하고 Swagger 연동 #78
- class-validator 기반 DTO 제거
- Zod 스키마를 단일 validation 소스로 사용
- Controller에서 ZodValidationPipe 적용
- Swagger request body schema를 Zod 기준으로 유지
